requirements.txt
aioredis==1.3.1
asgiref==3.3.1
astroid==2.4.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
autopep8==1.5.2
certifi==2020.4.5.1
cffi==1.14.0
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
cloudinary==1.20.0
constantly==15.1.0
coverage==5.1
cryptography==3.2
daphne==3.0.1
defusedxml==0.7.0rc1
Django==3.0.7
django-allauth==0.42.0
django-cors-headers==3.3.0
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-queryfields==1.0.0
djangorestframework-simplejwt==4.4.0
drf-jwt==1.15.1
ecdsa==0.15
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.8
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.2
oauthlib==3.1.0
Pillow==7.1.2
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyOpenSSL==20.0.1
python-dotenv==0.13.0
python-jose==3.1.0
python3-openid==3.1.0
pytz==2020.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
service-identity==18.1.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
toml==0.10.1
Twisted==20.3.0
txaio==20.12.1
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
zope.interface==5.2.0
